replacing lost app folder
