--- app/(app)/(auth)/login.js
@@ -1,2 +1,0 @@
-import LoginScreen from '../../../src/screens/LoginScreen';
-export default LoginScreen;
