--- app/(app)/(tabs)/index.tsx
@@ -1,2 +1,0 @@
-import HomeScreen from '../../../src/screens/HomeScreen';
-export default HomeScreen;
