--- app/(app)/(tabs)/log-workout.js
@@ -1,2 +1,0 @@
-import WorkoutLogScreen from '../../../src/screens/WorkoutLogScreen';
-export default WorkoutLogScreen;
