--- app/(app)/(tabs)/profile.js
@@ -1,2 +1,0 @@
-import ProfileScreen from '../../../src/screens/ProfileScreen';
-export default ProfileScreen;
