+++ app/(auth)/login.js
@@ -1,0 +1,4 @@
+import LoginScreen from '../../src/screens/LoginScreen';
+export default LoginScreen;
+
+// This file is used to export the LoginScreen component from the src/screens directory.
