+++ app/(tabs)/_layout.tsx
@@ -1,0 +1,27 @@
+import { Colors } from '@/constants/Colors';
+import { useColorScheme } from '@/hooks/useColorScheme';
+import { FontAwesome5 } from '@expo/vector-icons';
+import { Tabs } from 'expo-router';
+import React from 'react';
+
+export default function TabLayout() {
+  const colorScheme = useColorScheme();
+  return (
+    <Tabs
+      screenOptions={{
+        headerShown: false,
+        tabBarActiveTintColor: Colors[colorScheme ?? 'light'].tint,
+        tabBarInactiveTintColor: '#888',
+        tabBarStyle: {
+          backgroundColor: '#1a1a1a',
+          borderTopColor: '#333',
+        },
+      }}
+    >
+      <Tabs.Screen name="index" options={{ title: 'Home', tabBarIcon: ({ color }) => <FontAwesome5 name="home" size={24} color={color} /> }} />
+      <Tabs.Screen name="log-workout" options={{ title: 'Log Workout', tabBarIcon: ({ color }) => <FontAwesome5 name="dumbbell" size={24} color={color} /> }} />
+      <Tabs.Screen name="profile" options={{ title: 'Profile', tabBarIcon: ({ color }) => <FontAwesome5 name="user-alt" size={24} color={color} /> }} />
+      <Tabs.Screen name="journey-map" options={{ href: null }} />
+    </Tabs>
+  );
+}
