+++ app/(tabs)/journey-map.js
@@ -1,3 +1,2 @@
 import JourneyMapScreen from '../../src/screens/JourneyMapScreen';
-
 export default JourneyMapScreen;
