+++ app/(tabs)/log-workout.js
@@ -1,0 +1,7 @@
+import WorkoutLogScreen from '../../src/screens/WorkoutLogScreen';
+export default WorkoutLogScreen;
+
+// This file is used to export the WorkoutLogScreen component from the src/screens directory.
+// It allows the WorkoutLogScreen to be used in the app's navigation structure or as a standalone screen.
+// The WorkoutLogScreen component is typically used to log workouts, view workout history, and manage workout-related data.
+// It may include features such as adding new workouts, editing existing workouts, and
