+++ app/(tabs)/profile.js
@@ -1,0 +1,3 @@
+import ProfileScreen from '../../src/screens/ProfileScreen';
+export default ProfileScreen;
+
