+++ app/_layout.tsx
@@ -1,10 +1,30 @@
 import { useFonts } from 'expo-font';
-import { Slot, SplashScreen } from 'expo-router';
+import { Slot, SplashScreen, useRouter, useSegments } from 'expo-router';
 import React, { useEffect } from 'react';
-import { AuthProvider } from '../src/context/AuthContext';
+import { AuthProvider, useAuth } from '../src/context/AuthContext';
 
-SplashScreen.preventAutoHideAsync();
+// This component handles the redirection logic
+const InitialLayout = () => {
+    const { user, initializing } = useAuth();
+    const router = useRouter();
+    const segments = useSegments();
 
+    useEffect(() => {
+        if (initializing) return;
+
+        const inAuthGroup = segments[0] === '(auth)';
+
+        if (user && inAuthGroup) {
+            router.replace('/(tabs)');
+        } else if (!user && !inAuthGroup) {
+            router.replace('/(auth)/login');
+        }
+    }, [user, initializing, segments]);
+
+    return <Slot />;
+}
+
+// This is the main root layout
 export default function RootLayout() {
   const [loaded, error] = useFonts({
     'Cinzel': require('../assets/fonts/Cinzel-Regular.ttf'),
@@ -27,7 +47,7 @@ export default function RootLayout() {
 
   return (
       <AuthProvider>
-          <Slot />
+          <InitialLayout />
       </AuthProvider>
   );
 }
