--- app/app.js
@@ -1,7 +1,0 @@
-import HomeScreen from '../src/screens/HomeScreen';
-
-const App = () => {
-    return <HomeScreen />;
-};
-
-export default App;
